+++ packages/tldraw/src/lib/tools/SelectTool/childStates/Cropping/children/PointingCropHandle.ts
@@ -1,5 +1,5 @@
 import { StateNode, TLEventHandlers, TLImageShape, TLPointerEventInfo } from '@tldraw/editor'
-import { CursorTypeMap } from '../../PointingResizeHandle'
+import { CursorTypeMap } from '../../../select-helpers'
 
 type TLPointingCropHandleInfo = TLPointerEventInfo & {
 	target: 'selection'
