--- packages/tldraw/src/lib/tools/SelectTool/childStates/Cropping/crop-constants.ts
@@ -1,2 +1,0 @@
-/** @internal */
-export const MIN_CROP_SIZE = 8
